[SPARK-34943][BUILD] Upgrade flake8 to 3.8.0 or above in Jenkins
### What changes were proposed in this pull request?

Upgrade flake8 to 3.8.0 or above in Jenkins

### Why are the changes needed?

In flake8 < 3.8.0, F401 error occurs for imports in if statements when TYPE_CHECKING is True. However, TYPE_CHECKING is always False at runtime, so there is no need to treat it as an error in static analysis.

Since this behavior is fixed In flake8 >= 3.8.0, we should upgrade the flake8 installed in Jenkins to 3.8.0 or above. Otherwise, it occurs F401 error for several lines in pandas-on-PySpark that use TYPE_CHECKING

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Pass the CI

Closes #32749 from pingsutw/SPARK-34943.

Lead-authored-by: Kevin Su <pingsutw@gmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>