[SPARK-36870][SQL] Introduce INTERNAL_ERROR error class
### What changes were proposed in this pull request?

Adds the `INTERNAL_ERROR` error class and the `isInternalError` API to `SparkThrowable`.
Removes existing error classes that are internal-only and replaces them with `INTERNAL_ERROR`.

### Why are the changes needed?

Makes it easy for end-users to diagnose whether an error is an internal error. If an end-user encounters an internal error, it should be reported immediately. This also limits the number of error classes, making it easy to audit. We do not need high-quality error messages for internal errors, as they should not be exposed to the end-user.

### Does this PR introduce _any_ user-facing change?

Yes; this changes the error class in master.

### How was this patch tested?

Unit tests

Closes #34123 from karenfeng/internal-error-class.

Authored-by: Karen Feng <karen.feng@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>