[SPARK-36751][PYTHON][DOCS][FOLLOW-UP] Fix unexpected section title for Examples in docstring
### What changes were proposed in this pull request?

This PR is a minor followup of https://github.com/apache/spark/pull/33992 to fix the warnings during PySpark documentation build:

```
/.../spark/python/pyspark/sql/functions.py:docstring of pyspark.sql.functions.bit_length:19: WARNING: Unexpected section title or transition.

-------
/.../spark/python/pyspark/sql/functions.py:docstring of pyspark.sql.functions.octet_length:19: WARNING: Unexpected section title or transition.

-------
```

We should always have the same length of hyphens with the title.

### Why are the changes needed?

To remove warnings during the documentation build and show the HTML pages correctly.

### Does this PR introduce _any_ user-facing change?

This is not released yet, and only in master branch. So, no to end users.

### How was this patch tested?

Manually built the docs via `make clean html` at `python/docs` directory.

Closes #34196 from HyukjinKwon/SPARK-36751.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Kousuke Saruta <sarutak@oss.nttdata.com>