[SPARK-36713][PYTHON][DOCS] Document new syntax of type hints with index (pandas-on-Spark)
### What changes were proposed in this pull request?

This PR proposes to document the new syntax of type hints with index. Self-contained.

### Why are the changes needed?

To guide users about the new ways of typing to avoid creating default index.

### Does this PR introduce _any_ user-facing change?

Yes, it adds new sections in the pandas-on-Spark documentation.

### How was this patch tested?

Manually built the docs and verified the output HTMLs. Also manually ran the example codes.

![Screen Shot 2021-10-07 at 2 19 41 PM](https://user-images.githubusercontent.com/6477701/136324614-a9eafaa9-79b6-42fb-be65-ac43e12017b7.png)

![Screen Shot 2021-10-07 at 2 19 38 PM](https://user-images.githubusercontent.com/6477701/136324609-8da68d45-259e-441d-9226-b97fe7b0d63f.png)

Closes #34210 from HyukjinKwon/SPARK-36713.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>