[SPARK-37017][SQL] Reduce the scope of synchronized to prevent potential deadlock
### What changes were proposed in this pull request?

There is a `synchronize` block in `CatalogManager.currentNamespace` function.
This PR pulls `SessionCatalog.getCurrentDatabase` out from this `synchronize` block to prevent potential deadlock.

### Why are the changes needed?

In our case, we have implemented an external catalog, and there is a thread that directly calls SessionCatalog.getTempViewOrPermanentTableMetadata and holds the lock of SessionCatalog. It eventually goes into our external catalog, unfortunately, we then call some functions of SparkSession, e.g. sql. When it calls CatalogManager.currentNamespace, it tries to hold the lock of CatalogManager.

In the meantime, there are some query threads that execute sqls via DataFrame interface.

This is how deadlock occurs.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

use existed test.

Closes #34292 from chenzhx/bug-fix.

Authored-by: chenzhx <chen@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>